requirements.txt
cloudinary==1.36.0
fastapi==0.115.2
gradio_client==1.4.2
Pillow==9.0.1
Pillow==11.0.0
python-dotenv==1.0.1
Requests==2.32.3
twilio==8.2.2
